Added Recent Transactions. Added Funds Transfer User check
Account Creation, Depositing, Withdrawing, and Transfering Funds record the transaction, which can be viewed later. The balance after the transaction and the time the transaction took place are recorded as well. The 5 most recent transactions are displayed.

Transfer Funds will now check that the user ID to transfer to does exist.